Fixing bugs and glitches.
TODO: Get the fades working on CGB.

+++ src/inc/funcs.h
@@ -10,11 +10,11 @@
 
 extern void VBL_isr(void);
 extern void play(const hUGESong_t *Song);
-extern void show_dialog(uint16_t x, uint16_t y);
+extern void show_dialog(uint8_t x, uint8_t y);
 extern void win_print(unsigned char *text, uint8_t size);
 extern void scroll_sprites(uint8_t sprites, int8_t _dir, uint8_t vel);
 
-extern bool colliding(int16_t x, int16_t y, const unsigned char *map);
-extern bool canInteract(int16_t x, int16_t y, const unsigned char *map);
+extern bool colliding(uint8_t x, uint8_t y, const unsigned char *map);
+extern bool canInteract(uint8_t x, uint8_t y, const unsigned char *map);
 
 #endif
